introduce parameter `force_leaf` for TreeNode
That forces it to become a leaf node and defaults to False.

A TreeNode is considered a leaf if either this parameter is True or its
`load_child_keys` method returns None. This new parameter allows to declare a
node to be a leaf without the necessity of calling `load_child_keys`,
which is supposed to be used lazily.